Added stat tracking, another level
Currently configured for localhost, when you complete a level it sends
how long it took and how many moves to /flip/data via get. This is
supported in the most recent version of cosinegaming.com (it will
collect the data and display it with /flip/data/view).
Still no level can parallel the difficulty of the last level, which I
have commented because I don't want to figure out the par.